#2_acesso_ao_banco

+++ src/main/java/com/example/cardapio/food/FoodRequestDto.java
@@ -1,4 +1,4 @@
 package com.example.cardapio.food;
 
-public record FoodRequestDto(String title, String image, Integer price) {
+public record FoodRequestDto(String title, String image, float price) {
 }
